Fix deep status bug, and add promotion
The Chess::Rep package has a lot of issues, such as methods not actually
doing what the documentation says.  For now, getting around this by
manually modifying the internal data instead of calling the to_move
method in order to change the game state.  The status hashes should now
be correct when calculating the effects of a given move.