Allow strands to be built from a gene hash definition
Built a 'to_hash' method for the Gene class to export it's data in.
Started the Genetic Operator structure.  This is being built to allow an
entire set of strands to be processed by a given genetic operator.

The construction of the operator layout is what prompted the decision to
allow the strands to be constructed with a specific structure.  I
realized that passing around the objects themselves doesn't work when
you want to start running a reproduction algorithm on it.  There was no
way to guarantee that the structure you recieved would ever be the same.
Now IDs can be used to sort the list, allowing just the weights to be
passed around instead of enitre class data.